requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
requests==2.24.0
todoist-python==8.1.1
urllib3==1.25.9
